Create file structure and makefile
Created file structure. Source files go into `src`. Compiled code goes
into `build`. I've split the source code into a folder for each version
of the project, as a subdirectory of `src`.

In the makefile, I've only implemented this for the akosik one, so we
can start with tests for our implementation. In the end, each
implementation will have it's own build and test commands in the
makefile. Hopefully, this will serve us well.